Redis驱动新增选择DB和关闭断线重连

+++ src/Drivers/Redis.php
@@ -18,10 +18,13 @@ namespace easySwoole\Cache\Drivers;
 class Redis extends AbstractDriver
 {
 
-    protected $host   = '127.0.0.1';
-    protected $port   = 6379;
-    protected $auth   = null;
-    protected $prefix = 'cache:';
+    protected $host      = '127.0.0.1';
+    protected $port      = 6379;
+    protected $select    = 0;
+    protected $auth      = null;
+    protected $prefix    = 'cache:';
+    protected $reconnect = true;
+
     /* @var \Redis $handler */
     protected $handler = null;
 
@@ -37,75 +40,81 @@ class Redis extends AbstractDriver
      * 设置RedisHost
      * @param string $host
      * @author : evalor <master@evalor.cn>
-     * @return $this
      */
     function setHost($host)
     {
         $this->host = $host;
-        return $this;
     }
 
     /**
      * 设置RedisPort
      * @param string|int $port
      * @author : evalor <master@evalor.cn>
-     * @return $this
      */
     function setPort($port)
     {
         $this->port = intval($port);
-        return $this;
     }
 
     /**
      * 设置RedisAuthPass
      * @param string $password
      * @author : evalor <master@evalor.cn>
-     * @return $this
      */
     function setPassword($password)
     {
         $this->auth = (string)$password;
-        return $this;
     }
 
     /**
      * 设置缓存前缀
      * @param string $prefix
      * @author : evalor <master@evalor.cn>
-     * @return $this
      */
     function setPrefix($prefix)
     {
-        $this->prefix = $prefix;
-        return $this;
+        $this->prefix = (string)$prefix;
     }
 
     /**
-     * 连接到Redis
+     * 设置缓存使用的数据库序号
+     * @param $dbNum
      * @author : evalor <master@evalor.cn>
      */
-    protected function connect()
+    function setDatabase($dbNum)
     {
-        $this->handler = new \Redis();
-        $this->handler->connect($this->host, $this->port, 0);
-        if ($this->auth) $this->handler->auth($this->auth);
+        $this->select = intval($dbNum);
+    }
+
+    /**
+     * 设置是否开启断线重连
+     * @param $reconnect
+     * @author : evalor <master@evalor.cn>
+     */
+    function setReconnect($reconnect)
+    {
+        if (is_bool($reconnect)) $this->reconnect = $reconnect;
     }
 
     /**
      * 获取缓存
      * @param string $key
      * @param null $default
-     * @author : evalor <master@evalor.cn>
      * @return mixed|null
+     * @throws \Exception
+     * @author : evalor <master@evalor.cn>
      */
     public function get($key, $default = null)
     {
         try {
             $value = $this->handler->get($this->prefix . $key);
         } catch (\Exception $exception) {
-            $this->connect();
-            $value = $this->handler->get($this->prefix . $key);
+            if ($this->isBreak($exception)) {
+                $this->connect();
+                $value = $this->handler->get($this->prefix . $key);
+            } else {
+                throw $exception;
+            }
         }
 
         if (is_null($value) || false === $value) {
@@ -120,44 +129,53 @@ class Redis extends AbstractDriver
      * @param string $key
      * @param mixed $value
      * @param null $ttl
-     * @author : evalor <master@evalor.cn>
      * @return bool
+     * @throws \Exception
+     * @author : evalor <master@evalor.cn>
      */
     public function set($key, $value, $ttl = null)
     {
         if (!$ttl) $ttl = 0;
         $value = $this->pack($value);
-        if (is_int($ttl) && $ttl) {
-            try {
+        try {
+            if (is_int($ttl) && $ttl) {
                 $result = $this->handler->setex($this->prefix . $key, $ttl, $value);
-            } catch (\Exception $exception) {
-                $this->connect();
-                $result = $this->handler->setex($this->prefix . $key, $ttl, $value);
+            } else {
+                $result = $this->handler->set($this->prefix . $key, $value);
             }
-        } else {
-            try {
-                $result = $this->handler->set($this->prefix . $key, $value);
-            } catch (\Exception $exception) {
+            return $result;
+        } catch (\Exception $exception) {
+            if ($this->isBreak($exception)) {
                 $this->connect();
-                $result = $this->handler->set($this->prefix . $key, $value);
+                if (is_int($ttl) && $ttl) {
+                    $result = $this->handler->setex($this->prefix . $key, $ttl, $value);
+                } else {
+                    $result = $this->handler->set($this->prefix . $key, $value);
+                }
+                return $result;
+            } else {
+                throw $exception;
             }
         }
-        return $result;
     }
 
     /**
      * 删除缓存
      * @param string $key
-     * @author : evalor <master@evalor.cn>
      * @return bool
+     * @throws \Exception
+     * @author : evalor <master@evalor.cn>
      */
     public function delete($key)
     {
         try {
             $this->handler->delete($this->prefix . $key);
         } catch (\Exception $exception) {
-            $this->connect();
-            $this->handler->delete($this->prefix . $key);
+            if ($this->isBreak($exception)) {
+                $this->handler->delete($this->prefix . $key);
+            } else {
+                throw $exception;
+            }
         }
         return true;
     }
@@ -171,8 +189,43 @@ class Redis extends AbstractDriver
         try {
             $this->handler->delete($this->handler->keys($this->prefix . '*'));
         } catch (\Exception $exception) {
-            $this->connect();
-            $this->handler->delete($this->handler->keys($this->prefix . '*'));
+            if ($this->isBreak($exception)) {
+                $this->handler->delete($this->handler->keys($this->prefix . '*'));
+            } else {
+                throw $exception;
+            }
         }
+    }
+
+    /**
+     * 连接到Redis
+     * @author : evalor <master@evalor.cn>
+     */
+    protected function connect()
+    {
+        $this->handler = new \Redis();
+        $this->handler->connect($this->host, $this->port, 0);
+        if ($this->auth) $this->handler->auth($this->auth);
+        if ($this->select != 0) $this->handler->select($this->select);
+    }
+
+    /**
+     * 检测是否断线
+     * @param \Exception $e
+     * @author : evalor <master@evalor.cn>
+     * @return bool
+     */
+    protected function isBreak(\Exception $e)
+    {
+        if (!$this->reconnect) return false;
+        if (!$e instanceof \RedisException) return false;
+        $info = ['server went away'];
+        $error = $e->getMessage();
+        foreach ($info as $msg) {
+            if (false !== stripos($error, $msg)) {
+                return true;
+            }
+        }
+        return false;
     }
 }
